renamed package from controllers to controller
