+++ src/main/java/com/maks/durov/spring/intro/config/WebConfig.java
@@ -6,7 +6,7 @@ import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 
 @Configuration
 @EnableWebMvc
-@ComponentScan(basePackages = {"com.maks.durov.spring.intro.controllers"})
+@ComponentScan(basePackages = {"com.maks.durov.spring.intro.controller"})
 public class WebConfig {
 
 }
